update: modify code
see: #10
